move GaiaTileSet into its own folder
